Function refactor - pre hook change
